[Chore] modify chat send handler
