+++ client/src/components/chat/ChatForm.tsx
@@ -1,25 +1,13 @@
-import useHandleChat from 'hooks/useHandleChat';
 import React, { useState } from 'react';
-import { useParams } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-import { userState } from 'recoil/user/atoms';
 
-const ChatForm = () => {
-  const { chatId } = useParams();
-  const { userId } = useRecoilValue(userState);
-  const { sendHandler } = useHandleChat();
+const ChatForm = ({ onMessage }: { onMessage: (message: string) => void }) => {
   const [value, setValue] = useState<string>('');
 
   const handleKeyDown = (e: React.KeyboardEvent) => {
     if (e.code === 'Enter' && !e.shiftKey && !e.nativeEvent.isComposing) {
       e.preventDefault();
       const target = e.target as HTMLTextAreaElement;
-      const content = {
-        chatRoomId: chatId,
-        writerId: userId,
-        message: target.value,
-      };
-      sendHandler(content);
+      onMessage(target.value);
       setValue('');
     }
   };
