+++ client/src/pages/ChatPage.tsx
@@ -6,13 +6,15 @@ import { useGetChatData } from 'hooks/api/useGetChat';
 import useHandleChat from 'hooks/useHandleChat';
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import { useSetRecoilState } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { chatDataState } from 'recoil/chat/atoms';
+import { userState } from 'recoil/user/atoms';
 
 const ChatPage = () => {
   const { chatId } = useParams();
+  const { userId } = useRecoilValue(userState);
   const { data, isError, isLoading, isSuccess } = useGetChatData(chatId);
-  const { connectHandler, disconnectHandler } = useHandleChat();
+  const { connectHandler, disconnectHandler, sendHandler } = useHandleChat();
   const setChatData = useSetRecoilState(chatDataState);
 
   useEffect(() => {
@@ -25,17 +27,26 @@ const ChatPage = () => {
     if (isSuccess) {
       setChatData(data.data.chatList);
     }
-  }, [isSuccess, isError]);
+  }, [isSuccess, isError, data]);
 
   if (isError || isLoading) {
     return <></>;
   }
 
+  const handleMessage = (message: string) => {
+    const content = {
+      chatRoomId: chatId,
+      writerId: userId,
+      message,
+    };
+    sendHandler(content);
+  };
+
   return (
     <NoHeaderFooterLayout>
       <ChatUser user={data?.data.otherUserNickname} />
       <ChatMessages />
-      <ChatForm />
+      <ChatForm onMessage={handleMessage} />
     </NoHeaderFooterLayout>
   );
 };
